Refactor image namespace to card
