+++ src/components/App.js
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
 
-import ImageList from './ImageList';
+import CardList from './CardList';
 import Header from './Header';
 import { Archive, File, Upload, Check, X } from 'react-feather';
 import Text from './Text';
@@ -10,7 +10,7 @@ const { ipcRenderer } = require('electron');
 const App = () => {
     const capture = useRef(true);
 
-    const [ images, setImages ] = useState([]);
+    const [ cards, setCards ] = useState([]);
     const [ isDragging, setIsDragging ] = useState(false);
     const [ selected, setSelected ] = useState(0);
 
@@ -18,11 +18,11 @@ const App = () => {
         ipcRenderer.send('get-files');
 
         ipcRenderer.once('get-files-reply', (event, files) => {
-            setImages(files);
+            setCards(files);
         });
 
         ipcRenderer.on('new', (event, file) => {
-            setImages(previous => [ file, ...previous ]);
+            setCards(previous => [ file, ...previous ]);
         });
     }, []);
 
@@ -61,7 +61,7 @@ const App = () => {
     }, [isDragging]);
 
     const handleOnRemove = useCallback(path => {
-        setImages(current => {
+        setCards(current => {
             let r = current.filter(value => value !== path);
 
             if(r.length === 0) {
@@ -84,7 +84,7 @@ const App = () => {
 
             ipcRenderer.invoke('drop', paths)
                 .then(current => {
-                    setImages(previous => [ ...current, ...previous ]);
+                    setCards(previous => [ ...current, ...previous ]);
 
                     if(selected > -1) {
                         setSelected(0);
@@ -101,9 +101,9 @@ const App = () => {
 
     const handleOnKeyDown = useCallback(event => {
         if(capture.current) {
-            if(event.key === 'Enter' && images.length > 0) {
-                if(images[selected]) {
-                    ipcRenderer.send('copy', images[selected]);
+            if(event.key === 'Enter' && cards.length > 0) {
+                if(cards[selected]) {
+                    ipcRenderer.send('copy', cards[selected]);
                 }
             }
 
@@ -112,7 +112,7 @@ const App = () => {
             }
 
             if(event.key === 'Tab') {
-                if((selected < (images.length - 1)) && !event.shiftKey) {
+                if((selected < (cards.length - 1)) && !event.shiftKey) {
                     setSelected(current => {
                         return current + 1;
                     });
@@ -133,7 +133,7 @@ const App = () => {
         setTimeout(() => {
             capture.current = true;
         }, 100);
-    }, [selected, images]);
+    }, [selected, cards]);
 
     return (
         <div
@@ -163,7 +163,7 @@ const App = () => {
             <div className={`fixed rounded-lg transition-all top-0 bottom-0 right-0 left-0 bg-background-default z-20 pointer-events-none opacity-${isDragging ? '80' : '0'}`}/>
 
             {
-                (images.length === 0 && !isDragging) && <div
+                (cards.length === 0 && !isDragging) && <div
                     className={ 'transition-all text-center absolute top-10 right-0 left-0 bottom-0 flex pointer-events-none justify-center items-center text-background-accent text-background-accent'}>
                     <div>
                         <File size={ 180 }/>
@@ -177,7 +177,7 @@ const App = () => {
                 <span className={ 'ml-3' }>Archive</span>
             </Header>
 
-            <ImageList images={ images } onRemove={ path => handleOnRemove(path) } selected={selected} />
+            <CardList cards={ cards } onRemove={ path => handleOnRemove(path) } selected={selected} />
 
             <label
                 className={ 'text-background-default cursor-pointer fixed right-16 bottom-16 p-4 transition-all rounded-full bg-primary-default hover:bg-primary-accent' }>
